refactor(widget): split RichTextEditor into modular structure
Split the 3137-line rich_text_editor.rs into organized modules:

- block.rs: Block, BlockType, FormattedSpan
- link.rs: Link, ImageRef
- text_format.rs: TextFormat
- types.rs: ToolbarAction, EditorViewMode
- mod.rs: RichTextEditor widget with EditOp and DialogType

This improves code organization and maintainability while preserving
all functionality and exports.
